fix #307
