+++ irc/client_lookup_set.go
@@ -5,10 +5,8 @@
 package irc
 
 import (
-	"regexp"
 	"strings"
 	"sync"
-	"time"
 
 	"github.com/oragono/oragono/irc/caps"
 	"github.com/oragono/oragono/irc/modes"
@@ -305,135 +303,4 @@ func (clients *ClientManager) FindAll(userhost string) (set ClientSet) {
 	}
 
 	return set
-}
-
-//
-// usermask to regexp
-//
-
-//TODO(dan): move this over to generally using glob syntax instead?
-// kinda more expected in normal ban/etc masks, though regex is useful (probably as an extban?)
-
-type MaskInfo struct {
-	TimeCreated     time.Time
-	CreatorNickmask string
-	CreatorAccount  string
-}
-
-// UserMaskSet holds a set of client masks and lets you match  hostnames to them.
-type UserMaskSet struct {
-	sync.RWMutex
-	serialCacheUpdateMutex sync.Mutex
-	masks                  map[string]MaskInfo
-	regexp                 *regexp.Regexp
-}
-
-func NewUserMaskSet() *UserMaskSet {
-	return new(UserMaskSet)
-}
-
-// Add adds the given mask to this set.
-func (set *UserMaskSet) Add(mask, creatorNickmask, creatorAccount string) (maskAdded string, err error) {
-	casefoldedMask, err := CanonicalizeMaskWildcard(mask)
-	if err != nil {
-		return
-	}
-
-	set.serialCacheUpdateMutex.Lock()
-	defer set.serialCacheUpdateMutex.Unlock()
-
-	set.Lock()
-	if set.masks == nil {
-		set.masks = make(map[string]MaskInfo)
-	}
-	_, present := set.masks[casefoldedMask]
-	if !present {
-		maskAdded = casefoldedMask
-		set.masks[casefoldedMask] = MaskInfo{
-			TimeCreated:     time.Now().UTC(),
-			CreatorNickmask: creatorNickmask,
-			CreatorAccount:  creatorAccount,
-		}
-	}
-	set.Unlock()
-
-	if !present {
-		set.setRegexp()
-	}
-	return
-}
-
-// Remove removes the given mask from this set.
-func (set *UserMaskSet) Remove(mask string) (maskRemoved string, err error) {
-	mask, err = CanonicalizeMaskWildcard(mask)
-	if err != nil {
-		return
-	}
-
-	set.serialCacheUpdateMutex.Lock()
-	defer set.serialCacheUpdateMutex.Unlock()
-
-	set.Lock()
-	_, removed := set.masks[mask]
-	if removed {
-		maskRemoved = mask
-		delete(set.masks, mask)
-	}
-	set.Unlock()
-
-	if removed {
-		set.setRegexp()
-	}
-	return
-}
-
-func (set *UserMaskSet) SetMasks(masks map[string]MaskInfo) {
-	set.Lock()
-	set.masks = masks
-	set.Unlock()
-	set.setRegexp()
-}
-
-func (set *UserMaskSet) Masks() (result map[string]MaskInfo) {
-	set.RLock()
-	defer set.RUnlock()
-
-	result = make(map[string]MaskInfo, len(set.masks))
-	for mask, info := range set.masks {
-		result[mask] = info
-	}
-	return
-}
-
-// Match matches the given n!u@h.
-func (set *UserMaskSet) Match(userhost string) bool {
-	set.RLock()
-	regexp := set.regexp
-	set.RUnlock()
-
-	if regexp == nil {
-		return false
-	}
-	return regexp.MatchString(userhost)
-}
-
-func (set *UserMaskSet) Length() int {
-	set.RLock()
-	defer set.RUnlock()
-	return len(set.masks)
-}
-
-func (set *UserMaskSet) setRegexp() {
-	set.RLock()
-	maskExprs := make([]string, len(set.masks))
-	for mask := range set.masks {
-		maskExprs = append(maskExprs, mask)
-	}
-	set.RUnlock()
-
-	re, _ := utils.CompileMasks(maskExprs)
-
-	set.Lock()
-	set.regexp = re
-	set.Unlock()
 }
